+++ Plugins/UVOCBot.Plugins.Planetside/Objects/CensusQuery/MinimalCharacter.cs
@@ -1,0 +1,10 @@
+using DbgCensus.Core.Objects;
+
+namespace UVOCBot.Plugins.Planetside.Objects.CensusQuery;
+
+public record MinimalCharacter
+(
+    ulong CharacterID,
+    Name Name,
+    FactionDefinition FactionID
+);