+++ Plugins/UVOCBot.Plugins.Planetside/Services/BaseCachingPopulationService.cs
@@ -1,5 +1,6 @@
 ﻿using DbgCensus.Core.Objects;
 using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Extensions.Logging;
 using Remora.Results;
 using System.Threading;
 using System.Threading.Tasks;
@@ -12,18 +13,24 @@ namespace UVOCBot.Plugins.Planetside.Services;
 /// <inheritdoc cref="IPopulationService" />
 public abstract class BaseCachingPopulationService : IPopulationService
 {
+    private readonly ILogger<BaseCachingPopulationService> _logger;
     protected readonly IMemoryCache _cache;
 
-    protected BaseCachingPopulationService(IMemoryCache cache)
+    protected BaseCachingPopulationService(ILogger<BaseCachingPopulationService> logger, IMemoryCache cache)
     {
+        _logger = logger;
         _cache = cache;
     }
 
     /// <inheritdoc />
     public async Task<Result<IPopulation>> GetWorldPopulationAsync(ValidWorldDefinition world, CancellationToken ct = default, bool skipCacheRetrieval = false)
     {
-        if (!skipCacheRetrieval && _cache.TryGetValue(CacheKeyHelpers.GetPopulationKey((WorldDefinition)world), out IPopulation pop))
+        if (!skipCacheRetrieval &&
+            _cache.TryGetValue(CacheKeyHelpers.GetPopulationKey((WorldDefinition)world), out IPopulation pop))
             return Result<IPopulation>.FromSuccess(pop);
+
+        if (!skipCacheRetrieval)
+            _logger.LogWarning("Population was not retrieved from cache despite requesting it to be! Is the Census state worker not operating?");
 
         Result<IPopulation> popResult = await QueryPopulationAsync(world, ct).ConfigureAwait(false);
         if (!popResult.IsSuccess)